+++ page_analyzer/parsers/html.py
@@ -1,6 +1,4 @@
-from .models import Response
 from bs4 import BeautifulSoup
-import requests
 
 
 def get_seo_content(content):
@@ -15,17 +13,3 @@ def get_seo_content(content):
         content.title.string if content.title else None,
         meta_description['content'] if meta_description else None,
     )
-
-
-def get_response(url_name):
-    try:
-        resp = requests.get(url_name)
-        resp.raise_for_status()
-        return Response(
-            content=resp.content,
-            status_code=resp.status_code
-        )
-    except requests.exceptions.RequestException:
-        pass
-
-    return None
